requirements.txt
streamlit>=1.32.0
xgboost>=2.0.0
scikit-learn>=1.4.0
imbalanced-learn>=0.12.0
shap>=0.44.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.26.0
pdfplumber>=0.10.0
Pillow>=10.0.0
openpyxl>=3.1.0
pytesseract>=0.3.10
scipy>=1.12.0
